major upd 1
The following updates have been implemented: the introduction of custom components for buttons and badges, as well as the incorporation of compound classes. Additionally, modifications have been made to the Tailwind configuration to ensure that the applied rules include an !important flag, thereby mitigating the risk of any potential overrides.